Seperate tests based on file name
